# Code/Backend/document/apidoc.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/services": {
      "get": {
        "tags": [
          "service-controller"
        ],
        "operationId": "getAllServices",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListServiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/services/{id}": {
      "get": {
        "tags": [
          "service-controller"
        ],
        "operationId": "getServiceById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseServiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/services/{id}/load-balancers": {
      "get": {
        "tags": [
          "service-controller"
        ],
        "operationId": "getLoadBalancersByServiceId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListLoadBalancerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/services/{id}/group-modules": {
      "get": {
        "tags": [
          "service-controller"
        ],
        "operationId": "getGroupModulesByServiceId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListGroupModuleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/services/{id}/databases": {
      "get": {
        "tags": [
          "service-controller"
        ],
        "operationId": "getDatabasesByServiceId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListDatabaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/servers": {
      "get": {
        "tags": [
          "server-controller"
        ],
        "operationId": "getAllServers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListServerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/servers/{id}": {
      "get": {
        "tags": [
          "server-controller"
        ],
        "operationId": "getServerById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseServerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/servers/{id}/modules": {
      "get": {
        "tags": [
          "server-controller"
        ],
        "operationId": "getModulesByServerId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListModuleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/servers/{id}/databases": {
      "get": {
        "tags": [
          "server-controller"
        ],
        "operationId": "getDatabasesByServerId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListDatabaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/modules/{id}": {
      "get": {
        "tags": [
          "module-controller"
        ],
        "operationId": "getModuleById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseModuleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/modules/connections": {
      "get": {
        "tags": [
          "module-controller"
        ],
        "operationId": "getConnectionsBySourceModule_IdOrDestModule_Id",
        "parameters": [
          {
            "name": "sourceModuleId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "destModuleId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListConnectionResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiResponseListServiceResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceResponse"
            }
          }
        }
      },
      "ServiceResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "serviceCode": {
            "type": "string"
          },
          "serviceName": {
            "type": "string"
          }
        }
      },
      "ApiResponseServiceResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/ServiceResponse"
          }
        }
      },
      "ApiResponseListLoadBalancerResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoadBalancerResponse"
            }
          }
        }
      },
      "LoadBalancerResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "lbCode": {
            "type": "string"
          },
          "lbName": {
            "type": "string"
          },
          "ip": {
            "type": "string"
          },
          "port": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ApiResponseListGroupModuleResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupModuleResponse"
            }
          }
        }
      },
      "GroupModuleResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "groupModuleCode": {
            "type": "string"
          },
          "groupModuleName": {
            "type": "string"
          },
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleResponse"
            }
          }
        }
      },
      "ModuleResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "moduleCode": {
            "type": "string"
          },
          "moduleName": {
            "type": "string"
          }
        }
      },
      "ApiResponseListDatabaseResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DatabaseResponse"
            }
          }
        }
      },
      "DatabaseResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "databaseCode": {
            "type": "string"
          },
          "databaseName": {
            "type": "string"
          }
        }
      },
      "ApiResponseListServerResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServerResponse"
            }
          }
        }
      },
      "ServerResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "instanceIp": {
            "type": "string"
          }
        }
      },
      "ApiResponseServerResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/ServerResponse"
          }
        }
      },
      "ApiResponseListModuleResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleResponse"
            }
          }
        }
      },
      "ApiResponseModuleResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/ModuleResponse"
          }
        }
      },
      "ApiResponseListConnectionResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectionResponse"
            }
          }
        }
      },
      "ConnectionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "ipSource": {
            "type": "string"
          },
          "ipDest": {
            "type": "string"
          },
          "port": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    }
  }
}