# Code/Backend/src/main/java/vn/viettel/vdt_gd2/entity/dto/response/ConnectionResponse.java
package vn.viettel.vdt_gd2.entity.dto.response;

import lombok.*;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class ConnectionResponse {
    private Integer id;
    private String ipSource;
    private String ipDest;
    private Integer port;
    private Integer type;
}