# Code/Backend/src/main/java/vn/viettel/vdt_gd2/entity/dto/response/DatabaseResponse.java
package vn.viettel.vdt_gd2.entity.dto.response;

import lombok.*;
import lombok.experimental.FieldDefaults;

import java.util.List;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
@FieldDefaults(level = AccessLevel.PRIVATE)
public class DatabaseResponse {
    Integer id;
    String databaseCode;
    String databaseName;
}